Explicitly typecast uint16_t to int
Coverity CID 147132 & 147133

+++ src/plugins/select/cons_res/job_test.c
@@ -657,7 +657,8 @@ uint16_t _can_job_run_on_node(struct job_record *job_ptr, bitstr_t *core_map,
 				 * remove 1 cpu/core at a time. */
 				while ((cpus > 0) &&
 				       ((req_mem *
-					 (cpus * select_node_record[node_i].vpus))
+					 ((int) cpus *
+					  (int) select_node_record[node_i].vpus))
 					 > avail_mem))
 					cpus -= 1;
 			} else {
